Fixed CreateRequests migration

+++ db/migrate/20190524150500_create_beers.rb
@@ -4,7 +4,7 @@ class CreateBeers < ActiveRecord::Migration[5.2]
   def change
     create_table :beers do |t|
       t.string :name
-      t.integer :abv
+      t.float :abv
       t.text :description
 
       t.timestamps
